feat : textField Delegate를 이용한 입력 제한 및 모델 타입 변경

+++ OpenMarket/OpenMarket/Extension/UInt+extension.swift
@@ -1,5 +1,5 @@
 //
-//  Double+extension.swift
+//  UInt+extension.swift
 //  OpenMarket
 //
 //  Created by dudu, safari on 2022/05/14.
@@ -7,7 +7,7 @@
 
 import Foundation
 
-extension Double {
+extension UInt {
     func priceFormat(currency : String?) -> String? {
         let numberFormatter = NumberFormatter()
         numberFormatter.numberStyle = .decimal
